Add TODO for disabling hotkeys on fresh install
In order to use Cmd+Space for Launchbar instead of Spotlight, we disable
the relevant keyboard shortcut in `com.apple.symbolichotkeys`. However,
this plist uses a dictionary and the entry for shortcut `64` does not
exist on a fresh OS install until the checkbox is toggled in
Settings.app.

So far we just leave a comment explaining this, and a workaround.